Added dry_run flag and cleaned up path generation code.

+++ src/command.rs
@@ -16,7 +16,7 @@ pub fn unlink(options: &Options) -> Vec<Result<(), ()>> {
         .map(|package| {
             info!("Processing package {:?}", package);
             if let Ok(links) =
-                get_uninstall_paths::<_, &Path>(package, options.target.as_ref(), options.dotfiles)
+                get_paths::<_, &Path>(package, options.target.as_ref(), options.dotfiles, true)
             {
                 for link in links {
                     debug!("{:?}", link);
@@ -24,15 +24,19 @@ pub fn unlink(options: &Options) -> Vec<Result<(), ()>> {
                     if link.source.is_dir() {
                         if link.source.read_dir().unwrap().next().is_none() {
                             info!("Directory {:?} is empty, removing...", link.source);
-                            let res = remove_dir(link.source);
-                            debug!("remove_dir result {:?}", res);
+                            if !options.dry_run {
+                                let res = remove_dir(link.source);
+                                debug!("remove_dir result {:?}", res);
+                            }
                         }
                     } else if link.source.is_symlink()
                         && link.source.read_link().unwrap() == link.target
                     {
                         info!("Removing link: {:?} -> {:?}", link.source, link.target);
-                        let res = remove_file(link.source);
-                        debug!("remove_file result {:?}", res);
+                        if !options.dry_run {
+                            let res = remove_file(link.source);
+                            debug!("remove_file result {:?}", res);
+                        }
                     }
                 }
             } else {
@@ -55,14 +59,16 @@ pub fn link(options: &Options) -> Vec<Result<(), ()>> {
         .map(|package| {
             info!("Processing package {:?}", package);
             if let Ok(links) =
-                get_install_paths::<_, &Path>(package, options.target.as_ref(), options.dotfiles)
+                get_paths::<_, &Path>(package, options.target.as_ref(), options.dotfiles, false)
             {
                 for link in links {
                     if link.target.is_dir() {
                         if !link.source.exists() {
                             debug!("Making directory {:?}", link.source);
-                            let res = create_dir_all(link.source);
-                            debug!("create_dir_all result {:?}", res);
+                            if !options.dry_run {
+                                let res = create_dir_all(link.source);
+                                debug!("create_dir_all result {:?}", res);
+                            }
                         }
                     } else {
                         info!("Processing link: {:?} -> {:?}", link.source, link.target);
@@ -79,8 +85,10 @@ pub fn link(options: &Options) -> Vec<Result<(), ()>> {
                             }
                         } else if !link.source.exists() {
                             debug!("Making link {:?} -> {:?}", link.source, link.target);
-                            let res = symlink(link.target, link.source);
-                            debug!("symlink result {:?}", res);
+                            if !options.dry_run {
+                                let res = symlink(link.target, link.source);
+                                debug!("symlink result {:?}", res);
+                            }
                         }
                     }
                 }
@@ -109,10 +117,11 @@ where
     PathBuf::from(path)
 }
 
-fn get_install_paths<P, T>(
+fn get_paths<P, T>(
     package: P,
     target: T,
     map_dots: bool,
+    uninstall: bool,
 ) -> Result<Vec<Link>, walkdir::Error>
 where
     P: AsRef<Path>,
@@ -120,54 +129,23 @@ where
 {
     WalkDir::new(package)
         .min_depth(1)
+        .contents_first(uninstall)
         .into_iter()
         .map(|iter| {
             iter.map(|entry| {
                 let mut comp = entry.path().components();
                 comp.next();
-                let source = target.as_ref().join(comp.as_path());
+                let raw_source = target.as_ref().join(comp.as_path());
                 let mapped_source = match map_dots {
-                    true => map_path_dots(source),
-                    false => source,
+                    true => map_path_dots(raw_source),
+                    false => raw_source,
                 };
+                let source = shellexpand::full(mapped_source.to_str().expect("")).unwrap();
 
                 debug!("[TODO] Unwrapping canonicalize");
                 Link {
                     target: entry.path().canonicalize().unwrap().to_path_buf(),
-                    source: mapped_source,
-                }
-            })
-        })
-        .collect()
-}
-
-fn get_uninstall_paths<P, T>(
-    package: P,
-    target: T,
-    map_dots: bool,
-) -> Result<Vec<Link>, walkdir::Error>
-where
-    P: AsRef<Path>,
-    T: AsRef<Path>,
-{
-    WalkDir::new(package)
-        .min_depth(1)
-        .contents_first(true)
-        .into_iter()
-        .map(|iter| {
-            iter.map(|entry| {
-                let mut comp = entry.path().components();
-                comp.next();
-                let source = target.as_ref().join(comp.as_path());
-                let mapped_source = match map_dots {
-                    true => map_path_dots(source),
-                    false => source,
-                };
-
-                debug!("[TODO] Unwrapping canonicalize");
-                Link {
-                    target: entry.path().canonicalize().unwrap().to_path_buf(),
-                    source: mapped_source,
+                    source: source.into_owned().into(),
                 }
             })
         })
